+++ src/taskstore.py
@@ -12,10 +12,14 @@ from lootnika import (
 
 
 class TaskStore:
-    """
-    Use task logger
-    """
     def __init__(self, taskName: str, log: Logger, overwrite: bool = False):
+        """
+        Open taskstore or create new if it doesn't exist.\n
+        You must prepare taskstore before collect documents
+
+        :param log: Use task logger
+        :param overwrite: create new taskstore even if it exist
+        """
         self.log = log
         self.overwrite = overwrite
         self.cnx = self._create_task_store(taskName)
@@ -24,6 +28,7 @@ class TaskStore:
         """
         Connector creating local DB for each task
         to store information about seen documents
+
         :param taskName: will create taskName.db file
         :return: None on error
         """
@@ -74,6 +79,7 @@ class TaskStore:
         return cnx
 
     def prepare(self) -> bool:
+        """Mark all documents as old"""
         try:
             cur = self.cnx.cursor()
             cur.execute("UPDATE documents SET status='old'")
@@ -86,8 +92,9 @@ class TaskStore:
     def check_document(self, ref: str, docHash: str) -> int:
         """
         Check document for changes by hash
+
         :param ref: reference
-        :param docHash: cityHash64 hash from Document.get_hash()
+        :param docHash: cityHash64 from Document.get_hash()
         :return:
             operation status that can be:
                  0 - not changed\n
@@ -96,7 +103,7 @@ class TaskStore:
                 -1 - error\n
         """
         status = -1
-        self.log.info(f'Check object {ref}')
+        # self.log.debug(f'Check document {ref}')
         try:
             cur = self.cnx.cursor()
             cur.execute(f"SELECT hash FROM documents WHERE ref='{ref}'")
@@ -106,11 +113,11 @@ class TaskStore:
                     cur.executemany("INSERT INTO documents values(?,?,?)", [(ref, docHash, 'same')])
                     status = 0
                 else:
-                    self.log.info(f'Object has changed')
+                    self.log.info(f'Document {ref} has changed')
                     cur.executemany("INSERT INTO documents values(?,?,?)", [(ref, docHash, 'differ')])
                     status = 1
             else:
-                self.log.info(f'Object is new')
+                self.log.info(f'Document {ref} is new')
                 cur.executemany("INSERT INTO documents values(?,?,?)", [(ref, docHash, 'new')])
                 status = 2
 
@@ -147,6 +154,7 @@ class Document:
     def __init__(self, taskName: str, reference: str, fields: dict):
         assert isinstance(taskName, str)
         assert isinstance(reference, str)
+        assert isinstance(fields, dict)
 
         self.reference = reference
         self.raw = {
@@ -165,7 +173,13 @@ class Document:
             raise Exception(f"Missing the necessary @field@. Reference: {self.reference}")
 
     def get_hash(self) -> str:
-        return str(cityhash.CityHash64(orjson.dumps(self.raw, option=orjson.OPT_SORT_KEYS)))
+        """
+        calculate hash only for meta fields, not header
+        """
+        return str(cityhash.CityHash64(orjson.dumps(self.raw['fields'], option=orjson.OPT_SORT_KEYS)))
 
     def get_field(self, path: str):
+        """
+        using syntax from dpath library
+        """
         return dpath.get(self.raw, f'fields/{path}')